backend/requirements.txt
# FastAPI and server - LATEST COMPATIBLE VERSIONS 2024/2025
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==21.2.0
pydantic[email]==2.11.7
pydantic-settings==2.7.1
email-validator==2.1.2

# Authentication and security - STABLE VERSIONS
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==41.0.7

# Database - LATEST SUPABASE WITH STABLE DEPENDENCIES
supabase==2.17.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# AI and OpenAI - STABLE VERSIONS
openai==1.40.6
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Environment and configuration
python-dotenv==1.0.1

# Data analysis and processing - COMPATIBLE VERSIONS
scipy==1.11.4
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==14.0.2
lxml==4.9.3
PyYAML==6.0.1
chardet==5.2.0
fastavro==1.11.1

# File processing and validation
filetype==1.2.0

# Rate limiting and caching
slowapi==0.1.9
redis==5.0.1

# Utilities - STABLE VERSIONS
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Internal Scheduler for Cron Jobs - STABLE
APScheduler==3.10.4

# Logging and monitoring
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.35.0

# API Integration packages - STABLE COMPATIBLE
aiohttp==3.9.1
httpx==0.27.0
pycryptodome==3.19.0
requests-oauthlib==1.3.1

# SFTP integration
paramiko==3.4.0

# Development tools - STABLE VERSIONS
pytest==7.4.3
black==23.11.0
flake8==6.1.0 

# LangGraph for AI workflow orchestration  
langgraph==0.0.66
langchain==0.2.16
langchain-openai==0.1.25
langchain-core==0.2.40
langsmith==0.1.129

# Additional dependencies for enhanced template system
networkx==3.2.1
graphviz==0.20.1 